fix(ux): address PR review feedback
- Scope checkout try/catch to release-branch path only, so --rev
  errors propagate unchanged; include original error in message
- Fix incorrect comment about git branch sorting (alphabetical, not
  chronological)
- Extract shared formatArtifactConfigForError() into utils/strings.ts
  to eliminate YAML formatting duplication between publish.ts and
  github.ts
